Feature/mpt 16453 cleanup orphaned subpools (#322)
* Enhance teardown and pool management utilities

Added new utility functions for managing pools and data sources, including connecting, disconnecting, and deleting sub-pools by name. Updated API response types and enums for better environment handling. Improved global teardown logic to clean up orphaned Marketplace (Dev) sub-pools and reconnect data sources. Refactored test and helper code to use new types and functions.

* Refactor environment-based parent pool ID retrieval

Moved environment-specific parent pool ID logic from RestAPIRequest to a new utility function getEnvironmentParentPoolId in environment-util.ts. Added EParentPoolId enum for pool IDs and updated usages to improve maintainability and clarity. Also fixed minor issues in teardown-utils.ts and removed unnecessary debug logs.

* Refactor API header creation to use helper function

Introduced getBearerTokenHeader utility to centralize creation of authorization headers with Bearer tokens. Updated all relevant API request calls in restapi-request.ts and teardown-utils.ts to use this helper, improving code maintainability and reducing duplication.

* Update import-cloud-accounts-test.spec.ts

+++ e2etests/setup/global-teardown.ts
@@ -1,24 +1,41 @@
 import { request } from '@playwright/test';
 import { AuthRequest } from '../api-requests/auth-request';
 import { RestAPIRequest } from '../api-requests/restapi-request';
-import { cleanUpDirectoryIfEnabled, deletePolicies, deleteTestUsers } from '../utils/teardown-utils';
+import {
+  cleanUpDirectoryIfEnabled, connectDataSource,
+  deletePolicies,
+  deleteSubPoolsByName,
+  deleteTestUsers,
+  disconnectDataSource,
+  getDatasourceIdByNameViaOpsAPI,
+} from '../utils/teardown-utils';
 
 async function globalTeardown() {
-  const apiRequestContext = await request.newContext({
-    ignoreHTTPSErrors: true,
-    baseURL: process.env.BASE_URL,
-  });
-  const email = process.env.DEFAULT_USER_EMAIL;
-  const password = process.env.DEFAULT_USER_PASSWORD;
-  const authRequest = new AuthRequest(apiRequestContext);
-  const restAPIRequest = new RestAPIRequest(apiRequestContext);
-  const token = await authRequest.getAuthorizationToken(email, password);
 
-  await cleanUpDirectoryIfEnabled('./tests/downloads');
-  await deleteTestUsers(restAPIRequest, token);
-  await deletePolicies(restAPIRequest, token);
+  if (process.env.CLEAN_UP === 'true') {
+    const apiRequestContext = await request.newContext({
+      ignoreHTTPSErrors: true,
+      baseURL: process.env.BASE_URL,
+    });
+    const email = process.env.DEFAULT_USER_EMAIL;
+    const password = process.env.DEFAULT_USER_PASSWORD;
+    const authRequest = new AuthRequest(apiRequestContext);
+    const restAPIRequest = new RestAPIRequest(apiRequestContext);
+    const token = await authRequest.getAuthorizationToken(email, password);
 
-  await apiRequestContext.dispose();
+    await cleanUpDirectoryIfEnabled('./tests/downloads');
+    await deleteTestUsers(restAPIRequest, token);
+    await deletePolicies(restAPIRequest, token);
+
+    // clear down orphaned Marketplace (Dev) Sub-pools and reconnect data source
+    const dataSourceName = 'Marketplace (Dev)';
+    const marketplaceDevId = await getDatasourceIdByNameViaOpsAPI(restAPIRequest, dataSourceName);
+    await disconnectDataSource(restAPIRequest, token, marketplaceDevId);
+    await deleteSubPoolsByName(restAPIRequest, token, dataSourceName);
+    await connectDataSource(restAPIRequest, token, dataSourceName);
+
+    await apiRequestContext.dispose();
+  }
 }
 
 module.exports = globalTeardown;
